dispatch 시점과 스크립트 호출 시점 변경
