+++ src/Component/Map/Map.js
@@ -1,23 +1,16 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect} from 'react';
 import './Map.css';
-import {useDispatch, useSelector} from "react-redux";
-import {getInfo} from "../../Modules/getDatas";
+import {useSelector} from "react-redux";
 import NavBar from "../NavigationBar/NavBar";
 import KakaoMapScript from "./KakaoMapScript";
 
 
 export default function Map() {
-    const {data, loading, error} = useSelector(state => state.getDatas.datas);
-    const dispatch = useDispatch();
+    const {data} = useSelector(state => state.getDatas.datas);
 
-    useEffect(async () => {
-        await dispatch(getInfo());
+    useEffect(() => {
         KakaoMapScript(data);
-    }, [dispatch]);
-
-    if (loading) return <div>로딩중..</div>
-    if (error) return <div>에러발생</div>
-    if (!data) return <div>데이터가 없어요</div>;
+    })
 
     return (
         <div>
