+++ src/Modules/getDatas.js
@@ -19,7 +19,6 @@ export const getInfo = () => async dispatch => {
             payload: response.data
         })
         console.log('데이터를 가져왔습니다~');
-        // KakaoMapScript(response.data);
     } catch (e) {
         // 실패했을 떄
         dispatch({
