+++ src/Modules/index.js
@@ -1,7 +1,9 @@
 import {combineReducers} from "redux";
 import getDatas from "./getDatas";
+import keyword from "./keyword";
 
 const rootReducer = combineReducers({
+    keyword,
     getDatas
 });
